# backend-api/doc/openapiSpec.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Shopping Website API",
    "description": "REST API cho hệ thống thương mại điện tử – Node.js / Express / PostgreSQL",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "SuccessResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "message": {
            "type": "string"
          },
          "data": {
            "type": "object"
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false,
            "description": "Luôn là false khi có lỗi"
          },
          "message": {
            "type": "string",
            "description": "Nhãn bao quát: 'Bad request' (4xx) hoặc 'Internal server error' (5xx)",
            "example": "Bad request"
          },
          "error": {
            "type": "string",
            "description": "Chi tiết lỗi hiển thị cho client. 4xx: message thật từ service. 5xx: 'Hệ thống gặp sự cố, vui lòng thử lại sau.'",
            "example": "Chỉ còn 2 sản phẩm trong kho"
          },
          "stack": {
            "type": "string",
            "description": "Stack trace — chỉ có trong môi trường development",
            "nullable": true
          }
        },
        "required": ["success", "message", "error"]
      },
      "PaginationMeta": {
        "type": "object",
        "properties": {
          "page": {
            "type": "integer"
          },
          "limit": {
            "type": "integer"
          },
          "total": {
            "type": "integer"
          }
        }
      },
...
      "ProductImage": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "product_id": {
            "type": "integer"
          },
          "image_url": {
            "type": "string"
          },
          "is_thumbnail": {
            "type": "boolean"
          }
        }
      },
...
      "Order": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "order_code": {
            "type": "string"
          },
          "user_id": {
            "type": "integer"
          },
          "address_id": {
            "type": "integer",
            "nullable": true
          },
          "pickup_store_id": {
            "type": "integer",
            "nullable": true
          },
          "receiver_name": {
            "type": "string"
          },
          "receiver_phone": {
            "type": "string"
          },
          "shipping_address": {
            "type": "string"
          },
          "total_amount": {
            "type": "number"
          },
          "payment_method": {
            "type": "string"
          },
          "created_by_staff_id": {
            "type": "integer",
            "nullable": true
          },
          "created_by_staff_name": {
            "type": "string",
            "nullable": true,
            "description": "Tên staff tạo đơn (populated từ JOIN)"
          },
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "confirmed",
              "shipping",
              "completed",
              "cancelled"
            ]
          },
          "note": {
            "type": "string"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "shipping_fee": {
            "type": "number"
          }
        }
      },
...
      "OrderPreviewResult": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "product_id": {
                  "type": "integer"
                },
                "product_name": {
                  "type": "string"
                },
                "quantity": {
                  "type": "integer"
                },
                "unit_price": {
                  "type": "number"
                },
                "base_price": {
                  "type": "number"
                },
                "discount_amount": {
                  "type": "number"
                },
                "final_price": {
                  "type": "number"
                }
              }
            }
          },
          "shipping_fee": {
            "type": "number"
          },
          "total_base_amount": {
            "type": "number"
          },
          "total_discount_amount": {
            "type": "number"
          },
          "total_final_amount": {
            "type": "number"
          },
          "shipping_details": {
            "type": "object",
            "nullable": true,
            "properties": {
              "receiver_name": {
                "type": "string"
              },
              "receiver_phone": {
                "type": "string"
              },
              "shipping_address": {
                "type": "string"
              }
            }
          }
        }
      },
...
    },
    "parameters": {
      "IdParam": {
        "name": "id",
        "in": "path",
        "required": true,
        "schema": {
          "type": "integer"
        }
      },
      "PageParam": {
        "name": "page",
        "in": "query",
        "schema": {
          "type": "integer",
          "default": 1
        }
      },
      "LimitParam": {
        "name": "limit",
        "in": "query",
        "schema": {
          "type": "integer",
          "default": 10
        }
      },
...
    },
    "responses": {
      "Unauthorized": {
        "description": "Chưa xác thực — thiếu hoặc token không hợp lệ",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "success": false,
              "message": "Bad request",
              "error": "No token provided"
            }
          }
        }
      },
      "Forbidden": {
        "description": "Không có quyền truy cập",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "success": false,
              "message": "Bad request",
              "error": "Cart item not found"
            }
          }
        }
      },
      "NotFound": {
        "description": "Không tìm thấy tài nguyên",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "success": false,
              "message": "Bad request",
              "error": "Order not found"
            }
          }
        }
      },
      "BadRequest": {
        "description": "Dữ liệu đầu vào không hợp lệ",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "success": false,
              "message": "Bad request",
              "error": "product_id: Required"
            }
          }
        }
      },
      "InternalError": {
        "description": "Lỗi hệ thống",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "success": false,
              "message": "Internal server error",
              "error": "Hệ thống gặp sự cố, vui lòng thử lại sau."
            }
          }
        }
      }
    }
  },
  "paths": {
...
    "/api/product-images/{id}/thumbnail": {
      "patch": {
        "tags": ["Product Images"],
        "summary": "Đặt ảnh làm thumbnail (Admin)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/IdParam"
          }
        ],
        "responses": {
          "200": {
            "description": "Thumbnail đã cập nhật",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/SuccessResponse"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/ProductImage"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        },
        "operationId": "setThumbnail"
      }
    },
...
    "/api/orders/preview": {
      "post": {
        "tags": ["Orders"],
        "summary": "Xem trước đơn hàng",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["items"],
                "properties": {
                  "items": {
                    "type": "array",
                    "minItems": 1,
                    "items": {
                      "type": "object",
                      "properties": {
                        "product_id": {
                          "type": "integer"
                        },
                        "quantity": {
                          "type": "integer",
                          "minimum": 1
                        }
                      }
                    }
                  },
                  "address_id": {
                    "type": "integer"
                  },
                  "pickup_store_id": {
                    "type": "integer"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Kết quả tính toán đơn hàng",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/SuccessResponse"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/OrderPreviewResult"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        },
        "operationId": "previewOrder"
      }
    },
    "/api/orders": {
      "post": {
        "tags": ["Orders"],
        "summary": "Tạo đơn hàng",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["items"],
                "properties": {
                  "address_id": {
                    "type": "integer",
                    "description": "Bắt buộc nếu không có pickup_store_id"
                  },
                  "user_id": {
                    "type": "integer",
                    "description": "Chỉ dành cho admin/staff — ID của user được tạo đơn hộ"
                  },
                  "pickup_store_id": {
                    "type": "integer",
                    "description": "Bắt buộc nếu không có address_id"
                  },
                  "payment_method": {
                    "oneOf": [
                      {
                        "type": "string",
                        "enum": ["cod"]
                      },
                      {
                        "type": "integer",
                        "description": "ID phương thức thanh toán"
                      }
                    ]
                  },
                  "note": {
                    "type": "string"
                  },
                  "items": {
                    "type": "array",
                    "minItems": 1,
                    "items": {
                      "type": "object",
                      "required": ["product_id", "quantity"],
                      "properties": {
                        "product_id": {
                          "type": "integer"
                        },
                        "quantity": {
                          "type": "integer",
                          "minimum": 1
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Đơn hàng đã tạo",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/SuccessResponse"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/Order"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        },
        "operationId": "createOrder"
      },
      "get": {
        "tags": ["Orders"],
        "summary": "Lấy danh sách đơn hàng (Admin xem tất cả, User xem của mình)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/PageParam"
          },
          {
            "$ref": "#/components/parameters/LimitParam"
          },
          {
            "name": "status",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "pending",
                "confirmed",
                "shipping",
                "completed",
                "cancelled"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Danh sách đơn hàng",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/SuccessResponse"
                    },
                    {
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/Order"
                          }
                        },
                        "meta": {
                          "$ref": "#/components/schemas/PaginationMeta"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        },
        "operationId": "listOrders"
      }
    },
    "/api/orders/{id}": {
      "get": {
        "tags": ["Orders"],
        "summary": "Lấy chi tiết đơn hàng theo ID",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/IdParam"
          }
        ],
        "responses": {
          "200": {
            "description": "Chi tiết đơn hàng",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/SuccessResponse"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/Order"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        },
        "operationId": "getOrderById"
      },
      "put": {
        "tags": ["Orders"],
        "summary": "Cập nhật trạng thái đơn hàng (Admin)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/IdParam"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "status": {
                    "type": "string",
                    "enum": ["pending", "confirmed", "shipping", "completed"]
                  },
                  "note": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Đơn hàng đã cập nhật",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/SuccessResponse"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/Order"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        },
        "operationId": "updateOrder"
      },
      "delete": {
        "tags": ["Orders"],
        "summary": "Xóa đơn hàng (Admin)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/IdParam"
          }
        ],
        "responses": {
          "200": {
            "description": "Đã xóa",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        },
        "operationId": "deleteOrder"
      }
    },
    "/api/orders/{id}/cancel": {
      "post": {
        "tags": ["Orders"],
        "summary": "Hủy đơn hàng (User hoặc Admin)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/IdParam"
          }
        ],
        "responses": {
          "200": {
            "description": "Đơn hàng đã hủy",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          }
        },
        "operationId": "cancelOrder"
      }
    },
...
  }
}